# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Orb Task API
  version: 1.0.0
  description: |
    Machine-readable spec for the Orb task management API.
    Agents and integrations MUST consult this spec to know what is
    and is not possible before attempting any action.
  x-orb-agent-contract:
    description: >
      REST/API integration guidance for external agents and clients. The live
      conversational Orb prompt is governed by lib/orb-prompt.ts and assembled
      by app/actions/orb-converse.ts; this REST spec must not define or generate
      conversational prompt law.
    
    scope:
      default_project: ORB
      behavior: REST clients should pass explicit project filters when a request is project-scoped.
    
    valid_values:
      statuses: [open, in progress, on hold, closed]
      priorities:
        1: Urgent
        2: High
        4: Medium
        5: Low

  x-orb-internal-tools:
    query_repository:
      confidence: new
      description: 'Inspect the Orb source repository. Use this for questions about implementation, components, routes, configuration, documentation, or actual code behavior. On localhost, source=local reads the current working tree and source=production reads the current Vercel deployment. In production, only source=production is available. Access is restricted to Admin, Super Admin, and Developer roles.'
      input_schema:
        type: object
        properties:
          operation:
            type: string
...
          source:
...
          path:
...

...

...

...

...

...
servers:
...
security:
  - apiSecret: []
components:
  securitySchemes:
    apiSecret:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Raw secret string (not Bearer). Must match ORB_API_SECRET.
        API also requires ORB_API_ENABLED=true on the server.
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        todo_number:
          type: integer
          description: |
            Database-assigned current address number. Unique within the
            project and never reissued during that project's lifetime.
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, in progress, deferred, on hold, closed]
        priority_value:
          type: integer
          nullable: true
        resolution_notes:
          type: string
          nullable: true
        urls:
          type: array
          items:
            type: string
            format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
        due_at:
          type: string
          format: date-time
          description: "Due instant (timestamptz, ISO 8601 with offset). Interpreted for display in due_timezone."
          nullable: true
        due_timezone:
          type: string
          description: "IANA timezone the due time is expressed in (e.g. Pacific/Honolulu). Set whenever due_at is set."
          nullable: true
        reminder_lead_value:
          type: integer
          description: "Opt-in reminder lead amount (0-99); null with reminder_lead_unit means no reminder. 0 = at due time."
          nullable: true
        reminder_lead_unit:
          type: string
          enum: [minutes, hours, days, weeks, months]
          description: "Unit for reminder_lead_value. Months use calendar subtraction in due_timezone with the day clamped."
          nullable: true
    TaskCreate:
      type: object
      required: [product_code, title]
      properties:
        product_code:
          type: string
          description: Case-insensitive product code (e.g. ORB, HELM).
        title:
          type: string
        description:
          type: string
        priority_value:
          type: integer
        due_at:
          type: string
          description: "Optional due date/time. A wall-clock string (YYYY-MM-DDTHH:mm) is interpreted in due_timezone (or the project owner's timezone when omitted); an ISO string with offset is stored as-is."
          nullable: true
        due_timezone:
          type: string
          description: "IANA timezone for due_at (e.g. America/Vancouver). Defaults to the requester's current timezone in the app, or the project owner's stored timezone for headless API creates."
          nullable: true
        reminder_lead_value:
          type: integer
          minimum: 0
          maximum: 99
          description: "Opt-in reminder lead (set together with reminder_lead_unit). Omit for no reminder — reminders never affect the orb's mood. 0 = at due time."
          nullable: true
        reminder_lead_unit:
          type: string
          enum: [minutes, hours, days, weeks, months]
          nullable: true
    TaskUpdate:
      type: object
      description: |
        All fields are optional. Only provided fields are updated.
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in progress, deferred, on hold, closed]
          description: |
            Setting to a closed status (e.g. "closed") automatically sets closed_at to now.
            Setting to any other status automatically clears closed_at.
        priority_value:
          type: integer
        resolution_notes:
          type: string
        product_code:
          type: string
          description: |
            Move the task to a different project by providing the
            target project code (e.g. "HELM"). The database assigns the
            task a new, never-before-issued todo_number in the target project.
        urls:
          oneOf:
            - type: array
              items:
                type: string
            - type: string
              description: Newline-separated URLs, parsed server-side.
        due_at:
          type: string
          description: "Optional due date/time (wall-clock string interpreted in due_timezone, or ISO with offset), or null to clear — clearing also clears due_timezone and the reminder."
          nullable: true
        due_timezone:
          type: string
          description: "IANA timezone for due_at. Changing it alone re-labels the display zone without moving the stored instant."
          nullable: true
        reminder_lead_value:
          type: integer
          minimum: 0
          maximum: 99
          description: "Opt-in reminder lead; set together with reminder_lead_unit, or both null to remove the reminder. Any change re-arms a sent reminder."
          nullable: true
        reminder_lead_unit:
          type: string
          enum: [minutes, hours, days, weeks, months]
          nullable: true
...
paths:
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks for a product
      parameters:
        - name: product
          in: query
          required: true
          schema:
            type: string
          description: Product code (case-insensitive).
      responses:
        "200":
          description: Array of tasks ordered by todo_number ascending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          description: Missing product query parameter.
        "401":
          description: Invalid or missing Authorization header.
        "404":
          description: Product not found.
        "503":
          description: API disabled (ORB_API_ENABLED !== true).

    post:
      operationId: createTask
      summary: Create a new task
      x-orb-tool:
        name: create_todo
        confidence: well-tested
        description: |
          Create a new todo in a project. Defaults to the current project if product_code is omitted. Requires a title. Cannot create in a product that does not exist.
        parameter_overrides:
          product_code:
            description: "Target project code (e.g. ORB, HELM). Defaults to current project if omitted."
          title:
            description: "Task title. Required."
          priority_value:
            description: "Priority level (1=urgent, 2=high, etc). Omit if unknown."
          description:
            description: "Task description."
          due_at:
            description: "Optional due date/time as a wall-clock string, format YYYY-MM-DDTHH:mm — a reading in due_timezone. When the user names a time in a specific place (\"9am Tokyo time\"), pass that wall-clock here and the place's IANA zone in due_timezone."
          due_timezone:
            description: "IANA timezone the due time is expressed in (e.g. Asia/Tokyo, America/Vancouver). When the user names a city or zone, resolve it to the IANA identifier. Omit to use the user's current timezone."
          reminder_lead_value:
            description: "Reminder lead amount, 0-99, set together with reminder_lead_unit. Only when the user asks to be reminded (\"remind me a week before\" → 1 + weeks). Omit both otherwise — reminders are opt-in and never affect the orb's mood. 0 = at due time."
          reminder_lead_unit:
            description: "One of minutes, hours, days, weeks, months."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Missing product_code or title.
        "401":
          description: Unauthorized.
        "404":
          description: Product not found.
        "503":
          description: API disabled.
  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      operationId: updateTask
      summary: Update a task
      description: |
        Updates only the fields provided. Ignores soft-deleted tasks.

        NOT SUPPORTED:
        - Supplying or directly changing todo_number.
        - Changing created_at.

        To move a task between products, provide product_code. The task keeps
        its UUID and receives the target project's next database-owned,
        never-before-issued todo_number.
      x-orb-tool:
        name: update_todo
        confidence: well-tested
        description: |
          Update an existing todo by task code (e.g. ORB-73). Supports: title,
          status, priority, description, resolution_notes, urls. When the user
          supplies an exact task code already visible in their request or the
          BACKLOG, call update_todo directly; do not query_todos merely to
          reconfirm it. The server resolves and validates the current row.
          Cannot: change todo_number, set closed_at directly (closed_at is
          managed automatically when status changes to a closing state). To
          move a task between projects, use move_todo instead.
        parameters:
          - name: code
            in: custom
            required: true
            schema:
              type: string
            description: "Task code, e.g. \"ORB-73\"."
          - name: title_match
            in: custom
            required: false
            schema:
              type: string
            description: "Alternative: find task by partial title match."
        parameter_overrides:
          title:
            name: new_title
            description: "New title."
          status:
            name: new_status
            description: "New status value."
          priority_value:
            name: new_priority
            description: "New priority value."
          description:
            description: "Updated description."
          resolution_notes:
            description: "What was done to resolve the task. Populate when closing, not before."
          urls:
            description: "Related URLs."
          due_at:
            description: "Optional due date/time as a wall-clock string (YYYY-MM-DDTHH:mm) read in due_timezone, or null to clear (clearing also clears the timezone and reminder)."
          due_timezone:
            description: "IANA timezone for the due time (e.g. Asia/Tokyo). Resolve a spoken city or zone name to its IANA identifier; omit to keep the todo's stored zone."
          reminder_lead_value:
            description: "Reminder lead amount, 0-99, set together with reminder_lead_unit; null for both removes the reminder. Only when the user asks — reminders are opt-in and never affect the orb's mood."
          reminder_lead_unit:
            description: "One of minutes, hours, days, weeks, months."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized.
        "404":
          description: Task not found or already deleted.
        "503":
          description: API disabled.

    delete:
      operationId: deleteTask
      summary: Soft-delete a task
      description: Sets deleted_at timestamp. Does not destroy the row.
      x-orb-tool:
        name: delete_todo
        confidence: well-tested
        description: |
          Delete a todo by task code. If the user asks to delete all todos/tasks from a named project and BACKLOG already lists the matching task codes, call delete_todo once for each matching code immediately; do not query first just to confirm the visible list. The server may hold the deletion for confirmation depending on user preference; never claim deletion succeeded until the tool result confirms it.
        parameters:
          - name: code
            in: custom
            required: true
            schema:
              type: string
            description: "Task code, e.g. \"ORB-44\"."
      responses:
        "200":
          description: Soft-deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
        "401":
          description: Unauthorized.
        "503":
          description: API disabled.